+++ config/api.php
@@ -24,7 +24,7 @@ return [
                 'notes'             => OWC\OpenPub\Base\RestAPI\ItemFields\NotesField::class,
                 'synonyms'          => OWC\OpenPub\Base\RestAPI\ItemFields\SynonymsField::class,
                 'taxonomies'        => OWC\OpenPub\Base\RestAPI\ItemFields\TaxonomyField::class,
-                'escape_element'    => OWC\OpenPub\Base\RestAPI\ItemFields\EscapeElementField::class,
+                'escape_element'    => OWC\OpenPub\Base\RestAPI\ItemFields\EscapeElementField::class
             ],
         ],
         'theme'  => [