+++ config/taxonomies.php
@@ -10,15 +10,9 @@ return [
         'args'         => [
             'show_in_rest'      => true,
             'show_admin_column' => true,
-            'capabilities'      => [
-                //                'manage_terms' => 'manage_openpub_categories',
-                //                'edit_terms'   => 'manage_openpub_categories',
-                //                'delete_terms' => 'manage_openpub_categories',
-                //                'assign_terms' => 'edit_openpub_posts'
-            ]
+            'capabilities'      => []
         ],
         'names'        => [
-            # Override the base names used for labels:
             'singular' => _x('Audience', 'Taxonomy definition', 'openpub-base'),
             'plural'   => _x('Audiences', 'Taxonomy definition', 'openpub-base'),
             'slug'     => 'openpub-doelgroep'
@@ -30,15 +24,9 @@ return [
         'args'         => [
             'show_in_rest'      => true,
             'show_admin_column' => true,
-            'capabilities'      => [
-                //                'manage_terms' => 'manage_openpub_categories',
-                //                'edit_terms'   => 'manage_openpub_categories',
-                //                'delete_terms' => 'manage_openpub_categories',
-                //                'assign_terms' => 'edit_openpub_posts'
-            ]
+            'capabilities'      => []
         ],
         'names'        => [
-            # Override the base names used for labels:
             'singular' => _x('Type', 'Taxonomy definition', 'openpub-base'),
             'plural'   => _x('Types', 'Taxonomy definition', 'openpub-base'),
             'slug'     => 'openpub-type'
@@ -49,18 +37,23 @@ return [
         'args'         => [
             'show_in_rest'      => true,
             'show_admin_column' => true,
-            'capabilities'      => [
-                //                'manage_terms' => 'manage_openpub_categories',
-                //                'edit_terms'   => 'manage_openpub_categories',
-                //                'delete_terms' => 'manage_openpub_categories',
-                //                'assign_terms' => 'edit_openpub_posts'
-            ]
+            'capabilities'      => []
         ],
         'names'        => [
-            # Override the base names used for labels:
             'singular' => _x('Usage', 'Taxonomy definition', 'openpub-base'),
             'plural'   => _x('Usages', 'Taxonomy definition', 'openpub-base'),
             'slug'     => 'openpub-toepassing'
         ]
-    ]
+    ],
+    'openpub-show-on' => [
+        'object_types' => ['openpub-item'],
+        'args'         => [
+            'show_in_rest'      => false,
+            'show_admin_column' => true
+        ],
+        'names'        => [
+            'singular' => __('Show on', 'openpub-base'),
+            'plural'   => __('Show on', 'openpub-base')
+        ]
+    ],
 ];