+++ src/Base/Settings/SettingsPageOptions.php
@@ -47,13 +47,19 @@ class SettingsPageOptions
         return $this->settings['_owc_setting_use_escape_element'] ?? false;
     }
 
+    public function useShowOn(): bool
+    {
+        return $this->settings['_owc_setting_openpub_enable_show_on'] ?? false;
+    }
+
     public static function make(): self
     {
         $defaultSettings = [
             '_owc_setting_portal_url'               => '',
             '_owc_setting_portal_openpub_item_slug' => '',
             '_owc_setting_use_portal_url'           => 0,
-            '_owc_setting_use_escape_element'       => 0
+            '_owc_setting_use_escape_element'       => 0,
+            '_owc_setting_openpub_enable_show_on'   => false
         ];
 
         return new static(wp_parse_args(get_option('_owc_openpub_base_settings'), $defaultSettings));