+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version [2.0.5]
+
+### Feat
+
+-   Add 'show on' setting to openpub-settings.
+-   Add 'show on' setting in editor of openpub-item.
+-   Add filtering on 'show on' to items endpoint.
+
 ## Version [2.0.4]
 
 ### Feat